Added annotations and seprated 'OfflineData' and 'OfflinePreLearner' class into separate files.
